Make the local dev environment work an M1 Macbook
This commit sets the platform and architecture on which the flask,
vuejs, and redis containers will run to an x86 Linux using the
docker-compose `platform`  configuration. Without this flag, the
`flask` service fails to build on an M1 Macbook because Tensorflow,
which is required by smtag, only provides x86_64 binaries (M1 uses the
 ARM instruction set).
The standard Neo4j 4.1 image does not run on M1 Macbooks even with that
flag (not sure why, but probably due to the plugins). Instead, they
provide an experimental but working image for M1.

[Issue: #9]