Refactored Docmap view query to be more generic
The current Docmap query was hardcoded to return a single Docmap with a
first (review) step and, if present, a second (response) step. It didn't
handle cases with multiple Docmaps (e.g. one Docmap with reviews from
service A and then a publication Docmap at journal B) or of Docmaps with
more than two steps.

The database was already capable of holding this kind of information.
This commit only updates the view query to let it return unusual kinds of
Docmaps.