Removed duplicate logging during smart-tagging
smtag v2 uses tensorflow, which in turn uses abseil-py, which seems to
mess with the global logging config.
Preventing log propagation to the root logger seems to prevent this issue.
This was enabled for all relevant modules except if they were invoked
as scripts (e.g. `python -m sdg.sdneo`). This commit fixes that.