Adapted query to delete SDArticles for use in retagging
The query to delete all SDArticle nodes with a certain DOI, as well as
all SDFigure, SDPanel, and SDTag nodes connected to them, used to fail if
e.g. CondTag or Hypothesis nodes were already connected to it.
This is the case if the SDArticle had already been tagged and connected
to the main knowledge graph.
This commit allows us to trigger the retagging of any SDArticle even if
 it is already part of the knowledge graph by DETACH DELETE-ing.