Updated MECADOI ingestion to include significance sections
This is in preparation for generating summaries from these significance
sections. We could also extract them from the full text of the reviews,
which we are currently getting from hypothes.is, but we also wanted to
move towards extricating EEB from this dependency (and replacing it with
a dependency on MECADOI).