Added `participants` key with empty value to Docmaps
Some Docmaps Actions have natural participants, e.g. the authors are
participants in the `author-reply` Action. However, if there were no
obvious participants, we used to return the Action without the
`participants` key.
To better conform to how others are constructing/consuming Docmaps, this
commit adds a `participant` key to every Action, with an empty array
as value if there are no participants.