Fixed left-padding of main content
The left sidebar expands on hover and is otherwise 56px wide. However,
it forces a left-padding of the main content as if it was expanded, thus
wasting a lot of space.

This commit left-pads the main content to the width of the unexpanded
sidebar. The downside is that when it expands, some content is covered
up by the sidebar.